rip out fullpagedirective, make sidebar a child of the detail page as it always should have been
