+++ ui/src/app/cohort-review/cohort-review/cohort-review.component.css
@@ -1,42 +1,6 @@
-[appFullPage] {
-  position: relative;
-}
-
-.sidebar {
-  position: absolute;
-  top: 0;
-  right: 0px;
-  border: 1px solid rgb(86, 86, 86);
-  border-radius: 2px;
-  height: 100%;
-  z-index: 1;
+.full-page-div {
   background-color: rgb(250, 250, 250);
-  display: flex;
-}
-
-.sidebar-handle {
-  min-width: 24px;
-  background-color: rgb(230, 230, 230);
-}
-
-.cohort-review-main {
-  transition: margin .5s;
-  margin-right: 24px;
-}
-
-.cohort-review-main.sidebar-open {
-  margin-right: 324px;
-  padding-right: 1rem;
-}
-
-.sidebar-content {
-  width: 0;
-  transition: 0.5s;
-  overflow-x: hidden;
-  flex: 1;
-}
-
-.sidebar-content.sidebar-open {
-  margin-left: 0.25rem;
-  width: 300px;
+  padding: 1rem;
+  margin-bottom: 1rem;
+  position: relative;
 }
