+++ ui/src/app/cohort-review/cohort-review/cohort-review.component.html
@@ -1,21 +1,6 @@
-<div appFullPage>
+<div #fullPageDiv class="full-page-div">
   <!-- CohortReview main content container -->
-  <div class="cohort-review-main" [class.sidebar-open]="sidebarOpen">
-    <router-outlet></router-outlet>
-  </div>
-
-  <!-- Sidebar container -->
-  <div class="sidebar" *ngIf="state.participant$ | async">
-    <!-- Sidebar left column (always visible) -->
-    <div class="sidebar-handle" (click)="toggleSidebar()">
-      <clr-icon shape="angle-double" [dir]="angleDir" size="24"></clr-icon>
-    </div>
-    <!-- Sidebar content -->
-    <div class="sidebar-content" [class.sidebar-open]="sidebarOpen">
-      <app-sidebar-content>
-      </app-sidebar-content>
-    </div>
-  </div>
+  <router-outlet></router-outlet>
 
   <!-- Modals and other not in the flow stuff -->
   <!-- CohortReview first time creation modal -->
