+++ ui/src/app/cohort-review/cohort-review/cohort-review.component.ts
@@ -1,4 +1,11 @@
-import {Component, OnInit, ViewChild} from '@angular/core';
+import {
+  Component,
+  ElementRef,
+  HostListener,
+  OnInit,
+  Renderer2,
+  ViewChild,
+} from '@angular/core';
 import {ActivatedRoute} from '@angular/router';
 import {Subscription} from 'rxjs/Subscription';
 
@@ -7,6 +14,9 @@ import {ReviewStateService} from '../review-state.service';
 
 import {ReviewStatus} from 'generated';
 
+const pixel = (n: number) => `${n}px`;
+const ONE_REM = 24;  // value in pixels
+
 @Component({
   selector: 'app-cohort-review',
   templateUrl: './cohort-review.component.html',
@@ -14,10 +24,12 @@ import {ReviewStatus} from 'generated';
 })
 export class CohortReviewComponent implements OnInit {
   @ViewChild('createReviewModal') createReviewModal: CreateReviewComponent;
+  @ViewChild('fullPageDiv') fullPageDiv: ElementRef;
 
   constructor(
     private state: ReviewStateService,
     private route: ActivatedRoute,
+    private renderer: Renderer2,
   ) {}
 
   ngOnInit() {
@@ -29,21 +41,18 @@ export class CohortReviewComponent implements OnInit {
     if (review.reviewStatus === ReviewStatus.NONE) {
       this.createReviewModal.modal.open();
     }
+    this.updateWrapperDimensions();
   }
 
-  get angleDir() {
-    return this.sidebarOpen ? 'right' : 'left';
+  @HostListener('window:resize')
+  onResize() {
+    this.updateWrapperDimensions();
   }
 
-  get sidebarOpen() {
-    return this.state.sidebarOpen.getValue();
-  }
-
-  set sidebarOpen(value: boolean) {
-    this.state.sidebarOpen.next(value);
-  }
-
-  toggleSidebar() {
-    this.sidebarOpen = !this.sidebarOpen;
+  updateWrapperDimensions() {
+    const nativeEl = this.fullPageDiv.nativeElement;
+    const {top} = nativeEl.getBoundingClientRect();
+    const minHeight = pixel(window.innerHeight - top - ONE_REM);
+    this.renderer.setStyle(nativeEl, 'min-height', minHeight);
   }
 }
