+++ ui/src/app/cohort-review/participant-detail/participant-detail.component.css
@@ -1,7 +1,50 @@
+app-participant-detail {
+  position: relative;
+}
+
 .participant-browser-header {
     margin: 0px;
 }
 
 .participant-browser-navigation {
     display: flex;
+}
+
+.sidebar {
+  position: absolute;
+  top: 0;
+  right: 0px;
+  border: 1px solid rgb(86, 86, 86);
+  border-radius: 2px;
+  height: 100%;
+  z-index: 1;
+  background-color: rgb(250, 250, 250);
+  display: flex;
+}
+
+.sidebar-handle {
+  min-width: 24px;
+  background-color: rgb(230, 230, 230);
+}
+
+.main-content {
+  transition: margin .5s;
+  margin-right: 24px;
+}
+
+.main-content.sidebar-open {
+  margin-right: 324px;
+  padding-right: 1rem;
+}
+
+.sidebar-content {
+  width: 0;
+  transition: 0.5s;
+  overflow-x: hidden;
+  flex: 1;
+}
+
+.sidebar-content.sidebar-open {
+  margin-left: 0.25rem;
+  width: 300px;
 }
